refactor: move files with tests in separate folders
