+++ src/components/authButton/AuthButton.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { signOut } from 'firebase/auth';
 import { LoadingButton } from '@mui/lab';
-import { auth } from 'src/utils/firebase';
+import { auth } from 'src/utils/firebase/firebase';
 
 type AuthButtonProps = {
   title: string;
