+++ src/utils/authSubmit/authSubmit.ts
@@ -7,7 +7,7 @@ import {
   User,
 } from 'firebase/auth';
 import { AuthFormInputs } from 'src/types/AuthFormInputs';
-import { auth } from './firebase';
+import { auth } from '../firebase/firebase';
 
 export interface AuthSubmit extends AuthFormInputs {
   isSignIn: boolean;
